Updated serial conn interface implemented and tested.
This allows a message interface where the first byte tells the uController what type of data it is. Allows different controller modules in the backend to communicate with the uController on a "as-needed" basis, when an event/change occurs, without sending static data over serial port. Tested with Arduino connected.